{"schema":"decodex/commit/1","summary":"Split execution phase-goal orchestration","authority":"XY-1115"}
